test: extract schema matchers
