+++ src/schema.test.ts
@@ -1,40 +1,8 @@
 import * as TJS from "ts-json-schema-generator";
 import * as fs from "fs";
-import * as util from "util";
-import type { ValidateFunction } from "ajv";
 import Ajv from "ajv";
 import { makeTestCBNData } from "./data.test-helpers";
 import { expect, test } from "vitest";
-
-declare global {
-  namespace jest {
-    interface Matchers<R> {
-      toMatchSchema(validate: ValidateFunction): R;
-    }
-  }
-}
-
-expect.extend({
-  toMatchSchema(obj: any, schema: ValidateFunction) {
-    const valid = schema(obj);
-    const errors = schema.errors?.slice();
-    const filename = findFilename(obj, parentMap);
-    return {
-      pass: valid,
-      message: () => {
-        const errorMessages =
-          errors
-            ?.map(
-              (e) =>
-                `${e.instancePath} ${e.message}, but was ${util.inspect(e.data)}`,
-            )
-            .join("\n") ?? "";
-
-        return (filename ? `[File: ${filename}]\n` : "") + errorMessages;
-      },
-    };
-  },
-});
 
 const program = TJS.createGenerator({
   tsconfig: __dirname + "/../tsconfig.json",
@@ -70,34 +38,34 @@ const id = (x: any) => {
   if (x.om_terrain) return JSON.stringify(x.om_terrain);
 };
 
-const findFilename = (
-  obj: any,
-  parentMap: WeakMap<object, object | null>,
-): string | undefined => {
-  let current: any = obj;
-
-  while (current) {
-    if (current.__filename) return current.__filename;
-    current = parentMap.get(current) || null; // Move up to parent
-  }
-  return undefined;
-};
-
-// Create a parent tracking map before validation
-const parentMap = new WeakMap<object, object | null>();
-
-const buildParentMap = (obj: any, parent: any = null) => {
-  if (typeof obj !== "object" || obj === null) return;
-  parentMap.set(obj, parent);
-  for (const key in obj) {
-    if (typeof obj[key] === "object" && obj[key] !== null) {
-      buildParentMap(obj[key], obj);
-    }
-  }
-};
+// const findFilename = (
+//   obj: any,
+//   parentMap: WeakMap<object, object | null>,
+// ): string | undefined => {
+//   let current: any = obj;
+//
+//   while (current) {
+//     if (current.__filename) return current.__filename;
+//     current = parentMap.get(current) || null; // Move up to parent
+//   }
+//   return undefined;
+// };
+//
+// // Create a parent tracking map before validation
+// const parentMap = new WeakMap<object, object | null>();
+//
+// const buildParentMap = (obj: any, parent: any = null) => {
+//   if (typeof obj !== "object" || obj === null) return;
+//   parentMap.set(obj, parent);
+//   for (const key in obj) {
+//     if (typeof obj[key] === "object" && obj[key] !== null) {
+//       buildParentMap(obj[key], obj);
+//     }
+//   }
+// };
 
 // Build parent-child relationships
-buildParentMap(data.all());
+// buildParentMap(data.all());
 
 const all = data
   .all()
